+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Granular type inheritance
 - Default names to types
 - Functionality for making windows stay on top
+- Support for Caspar CGs HTML producer
 ### Changed
 - Some features have moved to the footer of the app window
 - Context menus now follow the color theme